src/requirements.txt
requests
pulsar-client==2.6.1
requests-toolbelt
pandas